Factor two type traits into RemoveReferenceWrapper.hpp
Factor out tt::remove_cvref_wrap and tt::remove_reference_wrapper
