+++ src/Domain/CoordinateMaps/DiscreteRotation.hpp
@@ -9,7 +9,7 @@
 
 #include "DataStructures/Tensor/TypeAliases.hpp"
 #include "Domain/OrientationMap.hpp"
-#include "Utilities/TypeTraits.hpp"
+#include "Utilities/TypeTraits/RemoveReferenceWrapper.hpp"
 
 /// \cond
 namespace PUP {
