+++ src/Domain/CoordinateMaps/Frustum.hpp
@@ -10,7 +10,7 @@
 
 #include "DataStructures/Tensor/TypeAliases.hpp"
 #include "Domain/OrientationMap.hpp"
-#include "Utilities/TypeTraits.hpp"
+#include "Utilities/TypeTraits/RemoveReferenceWrapper.hpp"
 
 /// \cond
 namespace PUP {
