+++ src/Domain/CoordinateMaps/ProductMapsTimeDep.tpp
@@ -18,7 +18,6 @@
 #include "Utilities/DereferenceWrapper.hpp"
 #include "Utilities/MakeWithValue.hpp"
 #include "Utilities/TMPL.hpp"
-#include "Utilities/TypeTraits.hpp"
 
 namespace domain {
 namespace CoordMapsTimeDependent {
