+++ src/Domain/CoordinateMaps/Rotation.hpp
@@ -12,11 +12,13 @@
 #include <limits>
 
 #include "DataStructures/Tensor/Tensor.hpp"
-#include "Utilities/TypeTraits.hpp"
+#include "Utilities/TypeTraits/RemoveReferenceWrapper.hpp"
 
+/// \cond
 namespace PUP {
 class er;
 }  // namespace PUP
+/// \endcond
 
 namespace domain {
 namespace CoordinateMaps {
